refactor(default-theme): Extract PageEdit component.
